requirements.txt
streamlit>=1.36
google-cloud-translate>=3.13
python-docx>=1.1
python-dotenv>=1.0
tqdm>=4.66
